Rewrite the tree-filter as a PHP script
Bash quoting issues in the inline tree-filter script were causing a ton of
problems that could not be detected until doing a run. As a result, it was
easier to rewrite that as a PHP script, passing it the arguments necessary to do
the work (essentially, any CLI arguments received, and a few things like root
directory path, PHP executable, etc).

The rewrite allowed removing several of the PHP utility scripts previously used.