Do not use a range!
Using a range caused the issues observed previously. Omitting the range did the
following:

- All tags were written correctly.
- All empty commits and merge commits were fully removed.
- Repo size went down dramatically (on zend-http, from > 210M to ~5M!).

Additionally, the script had an issue in that it saved the component's current
composer.json to a temp dir; unfortunately, when run in parallel, this meant
that it could get overwritten. This patch ensures they each have their own name.